requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-filetransfers==0.0.0
idna==2.8
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
